app/requirements.txt
Flask
Gunicorn
Werkzeug==2.2.2
python-dotenv
Flask-Cors
Flask-SQLAlchemy==2.5.0
SQLAlchemy==1.3.24
requests
psycopg2
Flask-Migrate==4.0.4
pandas
watchdog
sentry-sdk[flask]
python-telegram-bot
apscheduler
